Collapse .h and .cc files if possible in SimG4CMS/Calo
