+++ SimG4CMS/Calo/plugins/CaloSteppingAction.cc
@@ -1,34 +1,126 @@
-// to make hits in EB/EE/HC
-#include "SimG4CMS/Calo/interface/CaloSteppingAction.h"
-#include "SimG4Core/Notification/interface/G4TrackToParticleID.h"
+//#define EDM_ML_DEBUG
+//#define HcalNumberingTest
 
+// to make hits in EB/EE/HC
+
+#include "DataFormats/Common/interface/Handle.h"
 #include "DataFormats/HcalDetId/interface/HcalDetId.h"
-#include "SimDataFormats/CaloHit/interface/PCaloHit.h"
+#include "DataFormats/Math/interface/Point3D.h"
+
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/Framework/interface/ESHandle.h"
+#include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/Utilities/interface/isFinite.h"
+#include "FWCore/Utilities/interface/Exception.h"
+
+#include "Geometry/EcalCommonData/interface/EcalBarrelNumberingScheme.h"
+#include "Geometry/EcalCommonData/interface/EcalBaseNumber.h"
+#include "Geometry/EcalCommonData/interface/EcalEndcapNumberingScheme.h"
 #ifdef HcalNumberingTest
+#include "Geometry/HcalCommonData/interface/HcalNumberingFromDDD.h"
 #include "Geometry/HcalCommonData/interface/HcalDDDSimConstants.h"
 #include "Geometry/Records/interface/HcalSimNumberingRecord.h"
 #endif
 
-#include "FWCore/Framework/interface/EventSetup.h"
-#include "FWCore/Framework/interface/ESHandle.h"
-#include "FWCore/MessageLogger/interface/MessageLogger.h"
-#include "FWCore/Utilities/interface/isFinite.h"
-#include "FWCore/Utilities/interface/Exception.h"
+#include "SimDataFormats/CaloHit/interface/PCaloHit.h"
+#include "SimDataFormats/CaloHit/interface/PCaloHitContainer.h"
+#include "SimDataFormats/CaloHit/interface/PassiveHit.h"
+#include "SimDataFormats/SimHitMaker/interface/CaloSlaveSD.h"
 
+#include "SimG4CMS/Calo/interface/CaloGVHit.h"
+#include "SimG4CMS/Calo/interface/HcalNumberingScheme.h"
+#include "SimG4CMS/Calo/interface/HcalNumberingFromPS.h"
+
+#include "SimG4Core/Notification/interface/G4TrackToParticleID.h"
+#include "SimG4Core/Notification/interface/Observer.h"
+#include "SimG4Core/Notification/interface/BeginOfJob.h"
+#include "SimG4Core/Notification/interface/BeginOfRun.h"
+#include "SimG4Core/Notification/interface/BeginOfEvent.h"
+#include "SimG4Core/Notification/interface/EndOfEvent.h"
+#include "SimG4Core/Watcher/interface/SimProducer.h"
+#include "SimG4Core/Watcher/interface/SimWatcherFactory.h"
+
+#include "G4LogicalVolume.hh"
 #include "G4LogicalVolumeStore.hh"
 #include "G4NavigationHistory.hh"
 #include "G4ParticleTable.hh"
 #include "G4PhysicalVolumeStore.hh"
+#include "G4Region.hh"
 #include "G4RegionStore.hh"
+#include "G4Step.hh"
+#include "G4SystemOfUnits.hh"
+#include "G4Track.hh"
 #include "G4Trap.hh"
 #include "G4UnitsTable.hh"
-#include "G4SystemOfUnits.hh"
+#include "G4UserSteppingAction.hh"
+#include "G4VPhysicalVolume.hh"
+#include "G4VTouchable.hh"
 
+#include <algorithm>
 #include <cmath>
 #include <iostream>
 #include <iomanip>
+#include <string>
+#include <utility>
+#include <vector>
 
-//#define EDM_ML_DEBUG
+class CaloSteppingAction : public SimProducer,
+                           public Observer<const BeginOfJob *>,
+                           public Observer<const BeginOfRun *>,
+                           public Observer<const BeginOfEvent *>,
+                           public Observer<const EndOfEvent *>,
+                           public Observer<const G4Step *> {
+public:
+  CaloSteppingAction(const edm::ParameterSet &p);
+  ~CaloSteppingAction() override;
+
+  void produce(edm::Event &, const edm::EventSetup &) override;
+
+private:
+  void fillHits(edm::PCaloHitContainer &cc, int type);
+  void fillPassiveHits(edm::PassiveHitContainer &cc);
+  // observer classes
+  void update(const BeginOfJob *job) override;
+  void update(const BeginOfRun *run) override;
+  void update(const BeginOfEvent *evt) override;
+  void update(const G4Step *step) override;
+  void update(const EndOfEvent *evt) override;
+
+  void NaNTrap(const G4Step *) const;
+  uint32_t getDetIDHC(int det, int lay, int depth, const math::XYZVectorD &pos) const;
+  void fillHit(uint32_t id, double dE, double time, int primID, uint16_t depth, double em, int flag);
+  uint16_t getDepth(bool flag, double crystalDepth, double radl) const;
+  double curve_LY(double crystalLength, double crystalDepth) const;
+  double getBirkL3(double dE, double step, double chg, double dens) const;
+  double getBirkHC(double dE, double step, double chg, double dens) const;
+  void saveHits(int flag);
+
+  static const int nSD_ = 3;
+  std::unique_ptr<EcalBarrelNumberingScheme> ebNumberingScheme_;
+  std::unique_ptr<EcalEndcapNumberingScheme> eeNumberingScheme_;
+  std::unique_ptr<HcalNumberingFromPS> hcNumberingPS_;
+#ifdef HcalNumberingTest
+  std::unique_ptr<HcalNumberingFromDDD> hcNumbering_;
+#endif
+  std::unique_ptr<HcalNumberingScheme> hcNumberingScheme_;
+  std::unique_ptr<CaloSlaveSD> slave_[nSD_];
+
+  std::vector<std::string> nameEBSD_, nameEESD_, nameHCSD_;
+  std::vector<std::string> nameHitC_;
+  std::vector<const G4LogicalVolume *> volEBSD_, volEESD_, volHCSD_;
+  std::map<const G4LogicalVolume *, double> xtalMap_;
+  std::map<const G4LogicalVolume *, std::string> mapLV_;
+  int allSteps_, count_, eventID_;
+  double slopeLY_, birkC1EC_, birkSlopeEC_;
+  double birkCutEC_, birkC1HC_, birkC2HC_;
+  double birkC3HC_, timeSliceUnit_;
+  std::map<std::pair<int, CaloHitID>, CaloGVHit> hitMap_[nSD_];
+  typedef std::tuple<const G4LogicalVolume *, uint32_t, int, int, double, double, double, double, double, double, double>
+      PassiveData;
+  std::vector<PassiveData> store_;
+};
 
 CaloSteppingAction::CaloSteppingAction(const edm::ParameterSet& p) : count_(0) {
   edm::ParameterSet iC = p.getParameter<edm::ParameterSet>("CaloSteppingAction");
@@ -452,3 +544,8 @@ void CaloSteppingAction::saveHits(int type) {
                                     hit.second.getDepth());
   }
 }
+
+#include "SimG4Core/Watcher/interface/SimWatcherFactory.h"
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+
+DEFINE_SIMWATCHER(CaloSteppingAction);
