--- SimG4CMS/Calo/plugins/module.cc
@@ -1,7 +1,0 @@
-#include "SimG4CMS/Calo/interface/HcalTestAnalysis.h"
-#include "SimG4CMS/Calo/interface/CaloSteppingAction.h"
-#include "SimG4Core/Watcher/interface/SimWatcherFactory.h"
-#include "FWCore/PluginManager/interface/ModuleDef.h"
-
-DEFINE_SIMWATCHER(HcalTestAnalysis);
-DEFINE_SIMWATCHER(CaloSteppingAction);
